Clean up order by / group by query compilation
Unwinds a lot of spaghetti and redundancy. Both "group by" and
"order by" were using the same code, though the "nulls first/last"
behavior only mattered to order by, which littered a lot of the
implementation details with conditional checks on group.type
and/or item.nulls.

Some of the behavior was defined in the base QueryCompiler class,
overridden in dialect clients, and then passed back to the base class.

Clients were using the "format this item for inclusion in a group
by / order by element" to "fake" adding an additional column to
order by by returning something _other_ than what they were
expected to, and then the base method was checking if it
thought they did that and re-adding the thing that the subclass
was supposed to have formatted.

Now, order by and group by each have their own logic. Classes
declaratively choose their "order by nulls" behavior from a few
prebuilt functions, with the option to implement their own
behavior as an expression (which oracle uses, via IIF)