Add "wrapForExpressionList" to Formatter
The implementation of `wrap` pushes bindings onto the bindings
array when `isParameter` is true. It also wraps subqueries in
parenthesis when `isParameter` is true, and a bunch of deeply
nested code keeps checking the type of `value` to determine
which behavior to perform.

This new method is a start at decoupling those behaviors,
and allows us to wrap things "for use in an expression list"
(thus guaranteeing the wrapping-parenthesis) without making
changes that could impact the existing code.

Add tests for same